switch internal representation from list to map
ought to perform better on larger clusters. generally seems like a better fit anyway.
looking up a single node by UUID should be a fast operation whereas getting the list of all
nodes (which is now a little slower) usually only happens when dealing with an image upload
in which case it will be dwarfed by the actual image handling/resize stuff.

tracked in issue #30